Moved imgur functionality to a separate file.
